added interface Canvas, changed class Canvas to RectangleCanvas
